requirements.txt
openai>=1.0.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
matplotlib>=3.7.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
streamlit>=1.28.0
scikit-learn>=1.3.0
pandas>=2.0.0
